// content-directory-schemas/types/extrinsics/CreateClass.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type DefaultBoolean = boolean

/**
 * JSON schema to describe a new class for Joystream network
 */
export interface CreateClass {
  /**
   * Name of this class. Required property.
   */
  name: string
  /**
   * Description of this class.
   */
  description: string
  class_permissions?: {
    any_member?: DefaultBoolean
    entity_creation_blocked?: DefaultBoolean
    all_entity_property_values_locked?: DefaultBoolean
    maintainers?: number[]
  }
  maximum_entities_count: number
  default_entity_creation_voucher_upper_bound: number
}